Web GL

+++ Assignment 8/Assets/Scripts/CurrencyPickup.cs
@@ -1,13 +1,20 @@
+/* Julian Salgado
+ * CurrencyPickup.cs
+ * Assignment 8
+ * Subclass for currency pickup
+ */
+
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
 public class CurrencyPickup : Pickup
 {
-    [SerializeField] int value = 1;
+    [SerializeField] float value = 1;
 
     protected override void OnPickup(Player player)
     {
+        //player.coinCount += value;
         player.AddCoin(value);
     }
 }
